+++ src/commons/product_image_selection.rs
@@ -71,9 +71,14 @@ pub struct ResolvedProductImage {
 #[derive(Clone, Debug, Default, Deserialize, Eq, JsonSchema, PartialEq, Serialize)]
 #[serde(rename = "PascalCase")]
 #[derive(AsRefStr)]
+/// We default to `Always`, as we use floating tags for our release lines.
+/// This means the tag 23.4 starts of pointing to the same image 23.4.0 does, but switches to 23.4.1 after the releases of 23.4.1
+/// We let users always pull the latest version of the image, so that
+/// 1.) Users always get the latest version with security fixes
+/// 2.) We try our best that all pods use the same version and that not some nodes have an old version cached
 pub enum PullPolicy {
-    #[default]
     IfNotPresent,
+    #[default]
     Always,
     Never,
 }
@@ -144,7 +149,7 @@ mod tests {
             image: "docker.stackable.tech/stackable/superset:1.4.1-stackableoperator-version".to_string(),
             app_version_label: "1.4.1-stackableoperator-version".to_string(),
             product_version: "1.4.1".to_string(),
-            image_pull_policy: "IfNotPresent".to_string(),
+            image_pull_policy: "Always".to_string(),
             pull_secrets: None,
         }
     )]
@@ -158,7 +163,7 @@ mod tests {
             image: "docker.stackable.tech/stackable/superset:1.4.1-stackable2.1.0".to_string(),
             app_version_label: "1.4.1-stackable2.1.0".to_string(),
             product_version: "1.4.1".to_string(),
-            image_pull_policy: "IfNotPresent".to_string(),
+            image_pull_policy: "Always".to_string(),
             pull_secrets: None,
         }
     )]
@@ -173,7 +178,7 @@ mod tests {
             image: "my.corp/myteam/stackable/trino:1.4.1-stackable2.1.0".to_string(),
             app_version_label: "1.4.1-stackable2.1.0".to_string(),
             product_version: "1.4.1".to_string(),
-            image_pull_policy: "IfNotPresent".to_string(),
+            image_pull_policy: "Always".to_string(),
             pull_secrets: None,
         }
     )]
@@ -187,7 +192,7 @@ mod tests {
             image: "my.corp/myteam/stackable/superset".to_string(),
             app_version_label: "1.4.1-latest".to_string(),
             product_version: "1.4.1".to_string(),
-            image_pull_policy: "IfNotPresent".to_string(),
+            image_pull_policy: "Always".to_string(),
             pull_secrets: None,
         }
     )]
@@ -201,7 +206,7 @@ mod tests {
             image: "my.corp/myteam/stackable/superset:latest-and-greatest".to_string(),
             app_version_label: "1.4.1-latest-and-greatest".to_string(),
             product_version: "1.4.1".to_string(),
-            image_pull_policy: "IfNotPresent".to_string(),
+            image_pull_policy: "Always".to_string(),
             pull_secrets: None,
         }
     )]
@@ -216,7 +221,7 @@ mod tests {
             image: "my.corp/myteam/stackable/superset:latest-and-greatest".to_string(),
             app_version_label: "1.4.1-latest-and-greatest".to_string(),
             product_version: "1.4.1".to_string(),
-            image_pull_policy: "IfNotPresent".to_string(),
+            image_pull_policy: "Always".to_string(),
             pull_secrets: None,
         }
     )]