+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Changed
+
+- [BREAKING] ProductImageSelection now defaults `stackableVersion` to operator version ([#619]).
+- Default `pullPolicy` to operator `Always`` ([#619]).
+
+[#619]: https://github.com/stackabletech/operator-rs/pull/619
+
 ## [0.44.0] - 2023-07-13
 
 ### Added